This code defines a function count_prime_divisors(n) that takes an integer n as input and returns the number of prime divisors of n. It uses a helper function is_prime(n) to check if a number is prime.
The code then prompts the user to input 10 integers and stores them in a list called numbers. It then initializes two variables max_divisors and max_number to 0.

The code then iterates over the numbers list and for each number, it calculates the number of prime divisors using the count_prime_divisors(n) function. If the number of prime divisors is greater than the current max_divisors, it updates max_divisors and max_number with the current number and its number of prime divisors. If the number of prime divisors is equal to the current max_divisors, it updates max_number with the maximum of the current number and max_number.

Finally, the code prints the max_number and its number of prime divisors.

This code can be used to find the number with the maximum number of prime divisors among a given set of numbers.